float32 for predictions as well

+++ run_classifier_with_tfhub.py
@@ -131,9 +131,9 @@ def model_fn_builder(num_labels, learning_rate, num_train_steps,
 
         predictions = []
         for tensor in logits.numpy():
-          prediction = np.zeros(5)
+          prediction = np.zeros(5, dtype=np.float32)
           #threshold of 0.5
-          prediction[tensor>0.5] = 1
+          prediction[tensor>0.5] = 1.0
           predictions.append(prediction)
 
         predictions = np.array(predictions)
